our_DGMLA/requirement.txt
torch==2.2.1+cu121
torchvision==0.17.1+cu121
torch_geometric==2.5.0
torch_scatter==2.1.2+pt22cu121
torch_sparse==0.6.18+pt22cu121
numpy==1.26.4
Pillow==10.3.0
tqdm==4.66.4
warmup_scheduler==0.3
scikit-image==0.23.2
torchinfo==1.8.0
opencv-python==4.9.0.80